tools/tfc-orchestrator/pyproject.toml
[project]
name = "tfc-orchestrator"
version = "0.1.0"
description = "Autonomous Claude Code orchestrator for TFC-HMI — runs phases with fresh context, parallel stories, state tracking."
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0", "aiohttp>=3.9"]

[project.scripts]
tfc-orchestrator = "orchestrator.cli:main"

[tool.setuptools.packages.find]
include = ["orchestrator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
